requirements.txt
termcolor
groq
python-dotenv
huggingface-hub
praw